Unit test added for form text error, claim and citizen details
